+++ Desarrollo/SistemaVentasCatalogo/src/main/java/com/equipoanalisis/SistemaVentasCatalogo/entity/Producto.java
@@ -14,7 +14,7 @@ public class Producto {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id_producto")
-	private Long id_producto;
+	private int id_producto;
 
 	@Column(name = "id_tienda")
 	private Long id_tienda;
@@ -29,11 +29,11 @@ public class Producto {
 	private String foto;
 
 	
-	public Long getId_producto() {
+	public int getId_producto() {
 		return id_producto;
 	}
 
-	public void setId_producto(Long id_producto) {
+	public void setId_producto(int id_producto) {
 		this.id_producto = id_producto;
 	}
