+++ Desarrollo/SistemaVentasCatalogo/src/main/java/com/equipoanalisis/SistemaVentasCatalogo/entity/Rol.java
@@ -14,16 +14,16 @@ public class Rol {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id_rol")
-	private Long id_rol;
+	private int id_rol;
 	
 	@Column(name = "descripcion")
 	private String descripcion;
 
-	public Long getId_rol() {
+	public int getId_rol() {
 		return id_rol;
 	}
 
-	public void setId_rol(Long id_rol) {
+	public void setId_rol(int id_rol) {
 		this.id_rol = id_rol;
 	}
