+++ Desarrollo/SistemaVentasCatalogo/src/main/java/com/equipoanalisis/SistemaVentasCatalogo/entity/Tienda.java
@@ -14,7 +14,7 @@ public class Tienda {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id_tienda")
-	private Long id_tienda;
+	private int id_tienda;
 	
 	@Column(name = "id_categoria")
 	private Long id_categoria;
@@ -34,11 +34,11 @@ public class Tienda {
 	@Column(name = "cuota_envio")
 	private float cuota_envio;
 
-	public Long getId_tienda() {
+	public int getId_tienda() {
 		return id_tienda;
 	}
 
-	public void setId_tienda(Long id_tienda) {
+	public void setId_tienda(int id_tienda) {
 		this.id_tienda = id_tienda;
 	}
